icq: Move static data from instance to factory
